Spectrogram
This is a GUI for Spectrogram that I made as a term project for my Discrete Signal Processing course. It is designed with MATLAB AppDesigner. It allows the users to record sounds or upload sound files so that they can investigate the spectrogram of these sounds. Furthermore, it allows users to create signals and experiment with them in the same way.